% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap Standard Error of Correlation}
\usage{
bootstrap(var1, var2, k)
}
\arguments{
\item{var1}{A numeric vector.}

\item{var2}{A numeric vector.}

\item{k}{Number of bootstrap replications.}
}
\value{
A single numeric value: estimated standard error.
}
\description{
Uses bootstrapping to estimate the standard error of the correlation between two variables.
}
\examples{
x <- rnorm(100)
y <- 0.5 * x + rnorm(100, sd = 0.5)
bootstrap(x, y, k = 1000)

}
