% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_matrix.R
\name{bootstrap_matrix}
\alias{bootstrap_matrix}
\title{Bootstrap Matrix of Correlation Standard Errors}
\usage{
bootstrap_matrix(data, k)
}
\arguments{
\item{data}{A numeric data frame or matrix.}

\item{k}{Number of bootstrap replications.}
}
\value{
A symmetric matrix of standard errors, with variable names as row and column labels.
}
\description{
Computes a matrix of bootstrap-estimated standard errors for pairwise correlations.
}
\examples{
company_names <- c("Citi", "PFE", "GM")
k <- 500
set.seed(1052)
bootstrap_matrix(w_logret[, company_names], k)
}
