% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_regression.R
\name{bootstrap_regression}
\alias{bootstrap_regression}
\title{Bootstrap Standard Errors for Linear Regression Coefficients}
\usage{
bootstrap_regression(data, model_form, k)
}
\arguments{
\item{data}{A data frame containing the variables in the model.}

\item{model_form}{A formula specifying the linear regression model.}

\item{k}{An integer specifying the number of bootstrap samples to use.}
}
\value{
A numeric vector of bootstrap standard errors for each coefficient in the model.
}
\description{
Performs nonparametric bootstrap resampling to estimate the standard errors
of the coefficients from a linear regression model.
}
\examples{
bootstrap_se <- bootstrap_regression(w_logret, GM ~ Toyota + Ford, k = 1000)

}
