% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capm.R
\name{capm}
\alias{capm}
\title{Compute CAPM diagnostics: alpha, beta, Sharpe, Treynor (with delta method)}
\usage{
capm(stock_returns, market, rf)
}
\arguments{
\item{stock_returns}{A data frame or matrix of asset returns (each column is a stock).}

\item{market}{A numeric vector of market returns (e.g., S&P 500).}

\item{rf}{A numeric vector of risk-free rates (e.g., 3-month T-bill).}
}
\value{
A data frame with alpha, beta, their 95% CIs, Sharpe and Treynor ratios (with SEs)
}
\description{
Compute CAPM diagnostics: alpha, beta, Sharpe, Treynor (with delta method)
}
\examples{
stocks <- m_logret[, -1]
market <- m_sp500ret[, 2]
rf <- m_sp500ret[, 3]
capm_est <- capm(stocks, market, rf)

}
