% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqqplot.R
\name{multiqqplot}
\alias{multiqqplot}
\title{Q-Q Plots for Multiple Stocks}
\usage{
multiqqplot(data, company_names)
}
\arguments{
\item{data}{A data frame containing log return columns.}

\item{company_names}{A character vector of column names for companies to check.}
}
\value{
Q-Q plots for each stock in separate base R graphics panels.
}
\description{
Displays Q-Q plots of weekly log returns to assess normality for each stock.
}
\examples{
company_names <- c("Citi", "PFE", "GM")
multiqqplot(w_logret, company_names)
}
