% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot_comp.R
\name{qqplot_comp}
\alias{qqplot_comp}
\title{Q-Q Plot for Log Returns}
\usage{
qqplot_comp(data, company_name)
}
\arguments{
\item{data}{A data frame containing log return columns.}

\item{company_name}{A string: the column name of the company.}
}
\value{
A Q-Q plot.
}
\description{
Plots a Q-Q plot to assess normality of log returns for a single company.
}
\examples{
company_name <- c("Citi")
qqplot_comp(w_logret, company_name)

}
