% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Test Difference Between Two Coefficients in a Linear Model}
\usage{
t_test(model)
}
\arguments{
\item{model}{An object returned by `lm()`, representing a fitted linear model.}
}
\value{
A p-value indicating whether the first two coefficients are significantly different.
}
\description{
Performs a t-test on the first two coefficients of a linear model to check if they differ significantly.
}
\examples{
model <- lm(PFE ~ date_category, data = pfizer)
t_test(model)

}
